server refactor: move each service in its own file
