+++ server/app.js
@@ -15,9 +15,8 @@ var app = express();
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
-app.use(express.static(path.join(__dirname, 'public')));
-
 app.use('/', routes(sellerService, dispatcher));
+app.use(express.static(path.join(__dirname, 'public')));
 
 dispatcher.startBuying(1);
