+++ server/javascripts/services/index.js
@@ -1,0 +1,11 @@
+'use strict';
+
+var _ = require('lodash');
+
+var utils = require('../utils');
+
+exports.SellerService = require('./seller');
+exports.OrderService = require('./order');
+exports.Reduction = require('./reduction');
+exports.Dispatcher = require('./dispatcher');
+
